% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSqPeriodogram.R
\name{calcQp}
\alias{calcQp}
\title{calculate Qp}
\usage{
calcQp(values, varPer)
}
\arguments{
\item{values}{activity values (each value represents the measured activity in a minute)}

\item{varPer}{a period at which the chi-squared statistics is to be calculated}
}
\value{
a numeric of the calculated chi-squared statistics at the given varPer
}
\description{
calculate Qp
}
