% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSqPeriodogram.R
\name{chiSqPeriodogram}
\alias{chiSqPeriodogram}
\title{Calculate chi-square periodogram}
\usage{
chiSqPeriodogram(activityDF, periodRange = c(5, 35), res = 0.1)
}
\arguments{
\item{activityDF}{data frame containing time and activity values}

\item{periodRange}{a vector of two values indicating the range of periods to be tested}

\item{res}{time resolution for calculating chi-squared statistics}
}
\value{
data frame of two columns (dateTime (min), Qp value)
}
\description{
Calculate chi-square periodogram
}
