% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSqPeriodogramPlot.R
\name{chiSqPeriodogramPlot}
\alias{chiSqPeriodogramPlot}
\title{Draw a graph of chi-square periodogram}
\usage{
chiSqPeriodogramPlot(chiSqPrdgmDF)
}
\arguments{
\item{chiSqPrdgmDF}{data frame containing three column (testPerVec, Qp.act, Qp.sig)}
}
\value{
ggplot object
}
\description{
Draw a graph of chi-square periodogram
}
\examples{
oscillation <- sin(seq(0, 2 * pi * 10, by = 2 * pi / 1440))
oscillation.df <- data.frame(dateTime = 1:length(oscillation), value = oscillation)
chiSqPeriodogramPlot(chiSqPeriodogram(oscillation.df))

}
