% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiSqPeriodogramPlot.R
\name{getPeak}
\alias{getPeak}
\title{find the peak value from chi-squared periodogram}
\usage{
getPeak(x, y, z, p)
}
\arguments{
\item{x}{times at which each chi-squared statistics is calculated}

\item{y}{chi-squared statistics calculated from an activity data}

\item{z}{chi-squared statistics calculated from a null-hypothesis}

\item{p}{number of points to be used for fitting a quadratic function}
}
\value{
data frame with five numerics
}
\description{
find the peak value from chi-squared periodogram
}
